nexus-backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
